pyproject.toml
[project]
name = "backup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "gunicorn>=25.1.0",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "whitenoise>=6.12.0",
]
